Expose dumbo internals for testing
